Included ID field and constructor parameter for Product once again, made JDBCConnector work overall
Product has long ID as field again.
Field ID is long even though DB uses interger (serial in postgres) so
that it complies to the fpt.com.Product interface without having to
change it.
Made some structural changes to JDBCConnector:
I first tried to make it comply fpt.com.JDBCInterfact but dropped that
due to is not being a requirement.
Changed it so that the Connection con is openend in the constructor and
should be closed via JDBCConnector.close() after use.
Implemented some Exceptions.
And last but not least a little general restruction.
Looks like a lot but really isn't due to point 1.

Changes to JDBCMain not worth mentioning due to it's nature of being a
class only used for debugging.