Restructure JDBC ser and re-enable old ser strats
I introduced a new interface for the database IO called
"DatabaseStrategy" and made ViewShop differentiate between the two. The
Controller can then use one of Serializable- or Database-Strategy (the
second is way easier to work with).

IDs are now only generated when a SerializableStrat is selected and this
is done by the Controller. The DatabaseStrats check the ID for the
default value (-1) and only write those to the database to prevent
writing duplicates. Deleting is not supported.

Also, for that purpose, all products are now removed from the list when
the strategy is changed.